+++ scripts/s3/s3api/bucket_operations.py
@@ -23,7 +23,6 @@ from datetime import datetime, timedelta
 from time import perf_counter_ns
 
 from src.commons.constants import MIN_DURATION
-from src.commons.utils import corio_utils
 from src.libs import IAMClient
 from src.libs.s3api import S3Api
