+++ scripts/s3/s3api/copy_object_operations.py
@@ -25,7 +25,6 @@ from datetime import timedelta
 from time import perf_counter_ns
 
 from src.commons.constants import MIN_DURATION
-from src.commons.utils import corio_utils
 from src.libs.s3api import S3Api
