+++ scripts/s3/s3api/object_operations.py
@@ -24,7 +24,6 @@ from datetime import timedelta, datetime
 from time import perf_counter_ns
 
 from src.commons.constants import MIN_DURATION
-from src.commons.utils import corio_utils
 from src.libs.s3api import S3Api
