+++ src/commons/degrade_cluster.py
@@ -32,7 +32,7 @@ from munch import munchify
 from config import CLUSTER_CFG, S3_ENDPOINT
 from src.commons import constants as const
 from src.commons.exception import NoBucketExistsException
-from src.commons.utils.cluster_utils import ClusterServices
+from src.commons.utils.k8s import ClusterServices
 from src.libs.s3api.s3_bucket_ops import S3Bucket
 
 LOGGER = logging.getLogger(const.ROOT)
