+++ src/commons/report.py
@@ -27,9 +27,9 @@ from typing import Union
 import pandas as pd
 
 from src.commons.constants import ROOT
-from src.commons.utils.corio_utils import convert_size
-from src.commons.utils.corio_utils import get_report_file_path
-from src.commons.utils.corio_utils import monitor_sessions_iterations
+from src.commons.utils.utility import convert_size
+from src.commons.utils.utility import get_report_file_path
+from src.commons.utils.utility import monitor_sessions_iterations
 
 LOGGER = logging.getLogger(ROOT)
