+++ src/libs/s3api/s3_bucket_ops.py
@@ -24,7 +24,7 @@ import random
 import string
 import time
 
-from src.commons.utils.corio_utils import retries
+from src.commons.utils.utility import retries
 from src.libs.s3api.s3_restapi import S3RestApi
